feat: first cut of sscs-10294
