+++ src/test/java/uk/gov/hmcts/reform/sscs/helper/mapping/HearingsDurationMappingTest.java
@@ -1,0 +1,2 @@
+package uk.gov.hmcts.reform.sscs.helper.mapping;public class HearingsDurationMappingTest {
+}
